Simplifies: useMedia => useMatchMedia
